pyproject.toml
[project]
name = "telegram-marketplace"
version = "0.1.0"
description = "Telegram bot marketplace for excess produce deals"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.7",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "psycopg2-binary>=2.9",
    "alembic>=1.14",
    "SQLAlchemy>=2.0",
    "redis>=5.2",
    "stripe>=11.2",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "pytest-mock>=3.14",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-redis>=4.6",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing"
